v0.6.3 - Made suggestions: - Item ban bypass permission - remove "banned item cannot be moved" - permissions for every command - /rpg exp(needed)

+++ src/com/jabyftw/rpglv/ClassExecutor.java
@@ -5,10 +5,8 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.potion.PotionEffect;
-import org.bukkit.potion.PotionEffectType;
 
 /**
- *
  * @author Rafael
  */
 public class ClassExecutor implements CommandExecutor {
@@ -21,44 +19,54 @@ public class ClassExecutor implements CommandExecutor {
 
     @Override // /class (name/list/exit)
     public boolean onCommand(CommandSender sender, Command cmd, String lavel, String[] args) {
-        if (sender.hasPermission("rpglevel.join")) {
-            if (args.length < 1) {
+        if(sender.hasPermission("rpglevel.join")) {
+            if(args.length < 1) {
                 return false;
             } else {
-                if (args[0].equalsIgnoreCase("exit")) {
-                    if (sender instanceof Player) {
-                        Player p = (Player) sender;
-                        if (pl.players.containsKey(p)) {
-                            pl.sql.deletePlayer(p.getName().toLowerCase());
-                            p.setExp(0);
-                            p.setLevel(0);
-                            for (PotionEffect pet : p.getActivePotionEffects()) {
-                                p.removePotionEffect(pet.getType());
+                if(args[0].equalsIgnoreCase("exit")) {
+                    if(sender.hasPermission("rpglevel.exit")) {
+                        if(sender instanceof Player) {
+                            Player p = (Player) sender;
+                            if(pl.players.containsKey(p)) {
+                                pl.sql.deletePlayer(p.getName().toLowerCase());
+                                p.setExp(0);
+                                p.setLevel(0);
+                                for(PotionEffect pet : p.getActivePotionEffects()) {
+                                    p.removePotionEffect(pet.getType());
+                                }
+                                p.sendMessage(pl.getLang("youLeftClass").replaceAll("%name", pl.players.get(p).getClasse().getName()));
+                                pl.players.get(p).removeAllPermissions();
+                                pl.players.remove(p);
+                                return true;
+                            } else {
+                                p.sendMessage(pl.getLang("noClass"));
+                                return true;
                             }
-                            p.sendMessage(pl.getLang("youLeftClass").replaceAll("%name", pl.players.get(p).getClasse().getName()));
-                            pl.players.get(p).removeAllPermissions();
-                            pl.players.remove(p);
-                            return true;
                         } else {
-                            p.sendMessage(pl.getLang("noClass"));
+                            sender.sendMessage("Only ingame.");
                             return true;
                         }
                     } else {
-                        sender.sendMessage("Only ingame.");
+                        sender.sendMessage(pl.getLang("noPermission"));
                         return true;
                     }
-                } else if (args[0].equalsIgnoreCase("list")) {
-                    for (Classe c : pl.classes) {
-                        if (c.canJoin(sender)) {
-                            sender.sendMessage(pl.getLang("classList").replaceAll("%name", c.getName()).replaceAll("%exp", Double.toString(c.getExpNeeded(1))));
+                } else if(args[0].equalsIgnoreCase("list")) {
+                    if(sender.hasPermission("rpglevel.list")) {
+                        for(Classe c : pl.classes) {
+                            if(c.canJoin(sender)) {
+                                sender.sendMessage(pl.getLang("classList").replaceAll("%name", c.getName()).replaceAll("%exp", Double.toString(c.getExpNeeded(1))));
+                            }
                         }
+                        return true;
+                    } else {
+                        sender.sendMessage(pl.getLang("noPermission"));
+                        return true;
                     }
-                    return true;
                 } else {
-                    if (sender instanceof Player) {
+                    if(sender instanceof Player) {
                         String classe = args[0];
                         Classe c = pl.getClasse(classe);
-                        if (c.canJoin(sender)) {
+                        if(c.canJoin(sender)) {
                             c.addPlayer(((Player) sender));
                             return true;
                         } else {
